# frontend: use `mimirpb.Sample` custom JSON encoder func