--- a/pkg/frontend/querymiddleware/codec_test.go
+++ b/pkg/frontend/querymiddleware/codec_test.go
@@ -9,7 +9,6 @@ package querymiddleware
 import (
 	"bytes"
 	"context"
-	jsonstd "encoding/json"
 	"io"
 	"math/rand"
 	"net/http"
@@ -648,23 +647,6 @@ func BenchmarkPrometheusCodec_EncodeResponse(b *testing.B) {
 		_, err := PrometheusCodec.EncodeResponse(context.Background(), res)
 		require.NoError(b, err)
 	}
-}
-
-func TestPrometheusResponse_JSONEncode(t *testing.T) {
-	// Generate mocked response.
-	res := mockPrometheusResponse(1000, 1000)
-
-	resp, err := PrometheusCodec.EncodeResponse(context.Background(), res)
-	require.NoError(t, err)
-
-	out, err := io.ReadAll(resp.Body)
-	require.NoError(t, err)
-
-	stdOut, err := jsonstd.Marshal(res)
-	require.NoError(t, err)
-
-	// Compare against standard json output.
-	require.Equal(t, 0, bytes.Compare(out, stdOut))
 }
 
 func mockPrometheusResponse(numSeries, numSamplesPerSeries int) *PrometheusResponse {